+++ src/filename_utils.c
@@ -15,13 +15,13 @@ char *get_output_name(char *input_name) {
     int base_len = (int) (extension - input_name);
     char *suffix = "_processed.tif";
 
-    char *output = calloc(base_len + (int) strlen(suffix) + 1,
+    char *output = calloc((size_t) base_len + (size_t) strlen(suffix) + (size_t) 1,
             sizeof(char));
     if (output == NULL) {
         fprintf(stderr, "DEBUG: couldn't allocate space for new filename\n");
         return "";
     }
-    strncpy(output, input_name, base_len);
+    strncpy(output, input_name, (size_t) base_len);
     output[base_len] = '\0';
     strcat(output, suffix);
